requirements.txt
py2app==0.28
rumps==0.3.0
things.py==0.0.15
